fix(breakpoints): Breakpoints could get overwritten when multiple TS files map to a single JS file. (#285)
Maintain a map of source breakpoints per file. VSCode's
`setBreakpointRequest` is triggered per file whenever breakpoints are
added or removed. Since it does not provide all breakpoints for all
files, we need to maintain our own record of every breakpoint for every
file. This way, we have all inputs available when constructing the
complete list of BPs for a given generated file.