fix(source-maps): Use alternate source map bias only if the default returns null and the alternate does not. Remove as user option to reduce config confusion. (#303)
In an effort to reduce config options (and confusion), this removes the
sourceMapBias param from the launch config. The default source map bias
is 'leastUpperBound', which works with our published ts config in TS
Starter. This config is also used in the source-map unit tests. If
leastUpperBound fails, it will try greatestLowerBound but will not
switch to that as the preferred bias unless it succeeds.