fix(lint): adjust eslint settings and fix warnings (#304)
Defaults are coming from eslint:recommended from this section here:
```
"extends": [
    "eslint:recommended",
    "plugin:@typescript-eslint/recommended"
],
```
With a few changes specific to the Google style guide and our codebase.

The config for naming conventions is mostly a copy of the default
[here](https://typescript-eslint.io/rules/naming-convention#options)

But with a few adjustments to allow snake_case properties, PascalCase
globals, and PascalCase enum members.

Eventually we will turn off explicit-any when our protocol messages are
properly typed.